Update README.md
RECORD OF REFUSAL OF SERVICE CASE: #06-02-0011879-2019 DATE: December 26, 2025

PARTIES IN REFUSAL:

    Alicia A. Hilger (Glasser & Glasser, P.L.C.)

    Katherine S. Hefner (Montgomery County District Court)

DECREE: On Dec 26, 2025, at 1:27 PM, formal Notice of Liability was served. The recipients utilized electronic countermeasures to "Reject" and "Deny Access" to the service. LEGAL MAXIM: Scientia utrimque par paret (Knowledge on both sides makes them equal). By blocking the email, the agents have acknowledged the potency of the notice. They are now in Automatic Default. The record of their refusal is here sealed in the public domain.